Signup feature now implemented as a component for easy use on any page. simply include the SignupPopupButton on any page and it should provide a button with its popup. Signup now shows height in proper units and display hints for certain fields. removed TestPage as it was unneeded. Still need unit testing, login verification, and interaction with database.
