+++ client/src/components/Popup.css
@@ -17,14 +17,11 @@
   width: 100%;
   max-width: 640px;
   background: white;
-
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
 }
 
 .popup-inner .close-btn {
   position: absolute;
+  z-index: 1;
   top: 16px;
   right: 16px;
 }
