+++ client/src/components/SignupPopupButton.css
@@ -1,0 +1,16 @@
+.input-fields {
+  position: relative;
+  padding: 32px;
+  width: 100%;
+  max-width: 640px;
+  background: white;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.multi-field {
+  display: flex;
+  flex-direction: row;
+  gap: 16px;
+}
